+++ src/libtrx/include/libtrx/game/game_string.def
@@ -65,6 +65,7 @@ GS_DEFINE(OSD_AMBIGUOUS_INPUT_2, "Ambiguous input: %s and %s")
 GS_DEFINE(OSD_AMBIGUOUS_INPUT_3, "Ambiguous input: %s, %s, ...")
 GS_DEFINE(OSD_UI_ON, "UI enabled")
 GS_DEFINE(OSD_UI_OFF, "UI disabled")
+GS_DEFINE(OSD_GIVE_ITEM_ALL_KEYS, "Surprise! Every key item Lara needs is now in her backpack.")
 GS_DEFINE(OSD_COMPLETE_LEVEL, "Level complete!")
 GS_DEFINE(CONTROLS_DEFAULT_KEYS, "Default Keys")
 GS_DEFINE(CONTROLS_CUSTOM_1, "User Keys 1")
