+++ src/libtrx/include/libtrx/game/lara/cheat.h
@@ -2,7 +2,7 @@
 
 #include <stdint.h>
 
-extern bool Lara_Cheat_GiveAllKeys(void);
+bool Lara_Cheat_GiveAllKeys(void);
 extern bool Lara_Cheat_GiveAllGuns(void);
 extern bool Lara_Cheat_GiveAllItems(void);
 extern bool Lara_Cheat_EnterFlyMode(void);
