+++ src/tr1/game/lara/cheat.h
@@ -4,7 +4,6 @@
 
 bool Lara_Cheat_EnterFlyMode(void);
 bool Lara_Cheat_ExitFlyMode(void);
-bool Lara_Cheat_GiveAllKeys(void);
 bool Lara_Cheat_GiveAllGuns(void);
 bool Lara_Cheat_GiveAllItems(void);
 bool Lara_Cheat_OpenNearestDoor(void);
