+++ src/tr2/game/game_string.def
@@ -1,6 +1,5 @@
 GS_DEFINE(OSD_FLY_MODE_ON, "Fly mode enabled")
 GS_DEFINE(OSD_FLY_MODE_OFF, "Fly mode disabled")
-GS_DEFINE(OSD_GIVE_ITEM_ALL_KEYS, "Surprise! Every key item Lara needs is now in her backpack.")
 GS_DEFINE(OSD_GIVE_ITEM_ALL_GUNS, "Lock'n'load - Lara's armed to the teeth!")
 GS_DEFINE(OSD_GIVE_ITEM_CHEAT, "Lara's backpack just got way heavier!")
 GS_DEFINE(OSD_FLIPMAP_ON, "Flipmap set to ON")
