+++ src/tr2/game/lara/cheat.h
@@ -5,7 +5,6 @@
 bool Lara_Cheat_EnterFlyMode(void);
 bool Lara_Cheat_ExitFlyMode(void);
 void Lara_Cheat_GetStuff(void);
-bool Lara_Cheat_GiveAllKeys(void);
 bool Lara_Cheat_GiveAllGuns(void);
 bool Lara_Cheat_GiveAllItems(void);
 bool Lara_Cheat_OpenNearestDoor(void);
